Light cleanups to prepare for overflowdb-only generator
