+++ src/main/scala/io/shiftleft/fuzzyc2cpg/output/overflowdb/OverflowDbWriter.scala
@@ -4,9 +4,9 @@ import java.util.concurrent.BlockingQueue
 
 import better.files.File
 import io.shiftleft.codepropertygraph.cpgloading.{CpgLoader, ProtoToCpg}
-import overflowdb.OdbConfig
 import io.shiftleft.proto.cpg.Cpg.CpgStruct
 import org.slf4j.LoggerFactory
+import overflowdb.OdbConfig
 
 class OverflowDbWriter(outputPath: String, queue: BlockingQueue[CpgStruct.Builder]) extends Runnable {
